Corrections as per code review:
- Bumped branding to v4.11.
- Reverted VS version in the solution file.
- Removed brand override registry entry.
- Converted HealthCheck warning pop-up to info pop-up.
- Moved status report actions to a separate folder.
