+++ Jenkinsfile
@@ -30,7 +30,7 @@
  * SUCH DAMAGE.
  */
 
-def XENADMIN_BRANDING_TAG = 'v4.10'
+def XENADMIN_BRANDING_TAG = 'v4.11'
 
 @Library(['PacmanSharedLibrary', "xencenter-pipeline@v4.6"])
 import com.citrix.pipeline.xencenter.*
